extract types and constants from user model
