+++ src/models/constants.ts
@@ -1,0 +1,1 @@
+export const USER_KEY_PREFIX = "USER-";
